requirements.txt
pyfiglet==0.8.post0
PyJWT==2.6.0
termcolor==2.2.0
